06_CustomModal_Training/deployment/PyTorch/requirements.txt
torch==2.1.0
torchvision==0.16.0
numpy==1.26.0
pillow==10.0.1
matplotlib==3.8.0
tqdm==4.66.1
